First steps to multi dispatch with new dispatcher
This isn't just a case of wiring up what we had before to the dispatcher
mechanism, but reconsidering it in light of new opporunities. Also, this
is not actually used for most multiple dispatches yet - it's only for
those that are via qualified method calls, return value coercions that
end up needing to do a multiple dispatch, and so forth. Even in its
highly incomplete state, it does win back a decent number of tests.

Some differences with the current multi-dispatch and caching of it:

* We'll do a pre-filtering step to working out exactly which dispatches
  we can decide with dispatch program guards alone, and which need more
  consideration. Today, when we have things like `where` clauses, we
  give up entirely on caching. Under this new scheme, we'll keep track
  of which things, at the callsite, are valid candidates by nominal
  type, check those using the guards in the dispatch program, and then
  only have to consider the `where` clauses of those candidates that
  remain.
* If there's nothing that cares about concreteness, then we won't
  over-guard on that, as happens today because the cache mechanism
  doesn't really know what we care about.
* We can probably wire junction-y dispatches directly through to the
  junction auto-threader.